p-invoice/requirements.txt
playwright==1.51
requests
python-dotenv
schedule
pyyaml==6.0.1